Fixed some warnings
- Changing "char*" to "const char*" when a
  function can receive string literals
- Removed some unused variables and members
- Fixed some signed vs unsigned comparisons
- Added braces for safety on code like:
  if(...) if(...) ... else ...
  to make it:
  if(...) { if(...) ... else ... }
  otherwise the else is ambiguous
- Reordered some member initializations to
  fit the declaration order in the class
- OutDir must end with a backslash in VS

Reviewed-on: https://swiftshader-review.googlesource.com/3501
Tested-by: Alexis Hétu <sugoi@google.com>

+++ src/OpenGL/libGL/Program.h
@@ -111,11 +111,11 @@ namespace gl
 		void getAttachedShaders(GLsizei maxCount, GLsizei *count, GLuint *shaders);
 
 		void getActiveAttribute(GLuint index, GLsizei bufsize, GLsizei *length, GLint *size, GLenum *type, GLchar *name) const;
-		GLint getActiveAttributeCount() const;
+		size_t getActiveAttributeCount() const;
 		GLint getActiveAttributeMaxLength() const;
 
 		void getActiveUniform(GLuint index, GLsizei bufsize, GLsizei *length, GLint *size, GLenum *type, GLchar *name) const;
-		GLint getActiveUniformCount() const;
+		size_t getActiveUniformCount() const;
 		GLint getActiveUniformMaxLength() const;
 
 		void addRef();
